- mdmkolbe changed the definitions of u32, ush and uch in 	  include/csutil/zip.h to be uint32, uint16 and uint8 and 	  changed md5_word_t and md5_byte_t in include/csutil/csmd5.h 	  to be uint32 and uint8. 	- mdmkolbe changed the printf-like integer format strings to match 	  the following.  Formats for floats and strings were not changed. 		int i; unsigned u; 		printf("%d %X %u", i, i, u);
		short s; unsigned short us;
		long l; unsigned long l;
		printf("%hd %hu %ld %lu", s, us, l, ul);

		uint32 u32; size_t size; ptrdiff_t diff; int a, b; bool b;
		printf("%" PRId32 " %zu %td %td %d", u32, size, diff, &a - &b, (int) b);

		intptr_t ip; int *p;
		printf("%p %p", ip, p);


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@20784 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/import/map2cs/map.cpp
@@ -139,9 +139,9 @@ bool CMapFile::Read(const char* filename, const char* configfile)
   }
 
   printf("Map contains:\n");
-  printf("%lu Entites\n", (unsigned long)m_Entities.Length());
-  printf("%lu Brushes\n", (unsigned long)m_NumBrushes);
-  printf("%lu Unique planes\n", (unsigned long)m_Planes.Length());
+  printf("%zu Entites\n", m_Entities.Length());
+  printf("%zu Brushes\n", m_NumBrushes);
+  printf("%zu Unique planes\n", m_Planes.Length());
   return true;
 }
