+++ apps/tools/lighter/litobjsel.h
@@ -90,7 +90,7 @@ public:
     csRegExpMatchError rc = matcher.Match ("");
     if (rc == NoError || rc == NoMatch) return 0;
     static char error_string[100];
-    sprintf (error_string, "Regexp error %d", rc);
+    sprintf (error_string, "Regexp error %d", (int)rc);
     return error_string;
   }
   virtual bool SelectObject (iObject* obj)
@@ -211,7 +211,7 @@ public:
     csRegExpMatchError rc = matcher.Match ("");
     if (rc == NoError || rc == NoMatch) return 0;
     static char error_string[100];
-    sprintf (error_string, "Regexp error %d", rc);
+    sprintf (error_string, "Regexp error %d", (int)rc);
     return error_string;
   }
   virtual bool SelectObject (iObject* obj);
